+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changes to main, not yet committed to stable
 
 *Please add a brief description of any changes to be tested and any migrations to be performed here.*
--
+- Removed the normalization of display names (i.e. adding the " - 123-abc" suffix)
 
 ## 27 Mar 2024
 - Several changes to Synapse `homeserver.yaml` configuration, including a `ConfigChecker` module that